Add a TGRES_CLUSTER_DEBUG env setting for debug output, cluster is now less verbose by default.
Do not close the rpc endpoint, just let it die with the process.

Add a max-flushes-per-second config option to ratelimit the number
of database flushes so as to not overwhelm the db under high load.

Add ts reporting of aggregator and pacedmetric channel fill percent.

Add reporting of dispatcher channel fill percent and length.

Add a set-aside overrun queue. When the dispatcher channel is more
than half full, datapoints are set aside in the overrun queue. This
queue can be arpitrarily large. This way the dispatcher channel
does not need to be as big.

Pre-load all the known DSs into the cache on startup.

Report each flusher's channel fill percent.

Ensure that the node (host) name is part of any internal metric we
are tracking.

If the receiver is stopped, do not queue datapoints, just ignore them.

Fix some start/stop problems (mostly).

No DS flushes are requested from the worker loops, it is all now done
in the periodic flush only, which runs once a second.

The periodic flush will attempt to flush all the datapoints recently
accumulated, which may take several passes on a busy system. This
gives each datapoint about an equal chance of making it into the DB.

The ts update SQL now uses an "upsert" which eliminates the need for
pre-creating ts records (and makes ds creation faster).

Lots of tests are broken at the moment, will work on that next.