Introduce "watched ds" which are memory-resident (in addition to db) and can be used by the DSL without going to the database at all.
The watched DSs are kept in an LRU (thanks again Hashicorp).

Also fix FetchDataSources - using WITH ORDINALITY was wrong, it is not
an actual array index, but just an order and so results can be bogus
for sparse arrays.

A better SQL debug message for Series queries.

LoadRRAData can load an RRA with its data points from DB (needed for
watched DS).

A bunch of enhancements to the RRASeries since watched DS relies on
it heavily.

Fix Sum() to treat NaN as zero because this is what Graphite does,
apparently.